Refactor utils exports
