+++ app/javascript/controllers/snippets/editor.ts
@@ -1,7 +1,7 @@
 import { Controller } from '@hotwired/stimulus';
 
 import { debounce } from '../../utils/debounce';
-import debug from '../../utils/debug';
+import { debug } from '../../utils';
 
 const console = debug('app:javascript:controllers:snippets:editor');
