+++ app/javascript/controllers/snippets/tweet.ts
@@ -1,7 +1,7 @@
 import { Turbo } from '@hotwired/turbo-rails';
 import { Controller } from '@hotwired/stimulus';
 
-import debug from '../../utils/debug';
+import { debug } from '../../utils';
 
 const console = debug('app:javascript:controllers:snippets:tweet');
