+++ app/javascript/controllers/syntax-highlight/preview.js
@@ -1,6 +1,6 @@
 import { Controller } from '@hotwired/stimulus';
 
-import debug from '../../utils/debug';
+import { debug } from '../../utils';
 
 const console = debug('app:javascript:controllers:syntax-highlight:preview');
