+++ app/javascript/utils/debug.js
@@ -1,18 +1,18 @@
-import debug from 'debug';
+import _debug from 'debug';
 
-export default function (namespace) {
+export const debug = (namespace) => {
   return {
-    enable: debug.enable.bind(debug),
-    log: debug(namespace),
+    enable: _debug.enable.bind(debug),
+    log: _debug(namespace),
     error: (() => {
-      const error = debug(namespace);
+      const error = _debug(namespace);
       error.log = console.error.bind(console);
       return error;
     })(),
     warn: (() => {
-      const warn = debug(namespace);
+      const warn = _debug(namespace);
       warn.log = console.warn.bind(console);
       return warn;
     })(),
   };
-}
+};
