+++ app/javascript/utils/plausible.js
@@ -1,0 +1,5 @@
+export const plausible =
+  window.plausible ||
+  function () {
+    ' ';
+  };
